feat: add base node class
